+++ domain/learningandworkprogress/src/test/kotlin/uk/gov/justice/digital/hmpps/domain/learningandworkprogress/review/service/ReviewServiceTest.kt
@@ -14,12 +14,7 @@ import org.mockito.kotlin.verify
 import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.review.ReviewScheduleNotFoundException
 import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.review.aValidCompletedReview
 import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.review.aValidReviewSchedule
-import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.review.aValidReviewScheduleHistory
-import java.time.Instant
-import java.time.LocalDate
-import java.time.temporal.ChronoUnit.DAYS
-import java.time.temporal.ChronoUnit.MINUTES
-import java.util.UUID
+import uk.gov.justice.digital.hmpps.domain.randomValidPrisonNumber
 
 @ExtendWith(MockitoExtension::class)
 class ReviewServiceTest {
@@ -36,8 +31,7 @@ class ReviewServiceTest {
   private lateinit var reviewSchedulePersistenceAdapter: ReviewSchedulePersistenceAdapter
 
   companion object {
-    private const val PRISON_NUMBER = "A1234AB"
-    private val TODAY = LocalDate.now()
+    private val PRISON_NUMBER = randomValidPrisonNumber()
   }
 
   @Nested
@@ -120,85 +114,6 @@ class ReviewServiceTest {
       // Then
       assertThat(actual).isEqualTo(expected)
       verify(reviewPersistenceAdapter).getCompletedReviews(PRISON_NUMBER)
-    }
-  }
-
-  @Nested
-  inner class GetReviewSchedulesForPrisoner {
-    @Test
-    fun `should get review schedules for prisoner, sorted by last updated and version`() {
-      // Given
-      val now = Instant.now()
-      val reviewSchedule1Reference = UUID.randomUUID()
-      val reviewSchedule2Reference = UUID.randomUUID()
-      val reviewSchedule3Reference = UUID.randomUUID()
-
-      given(reviewSchedulePersistenceAdapter.getReviewScheduleHistory(any())).willReturn(
-        listOf(
-          aValidReviewScheduleHistory(
-            reference = reviewSchedule3Reference,
-            version = 2,
-            lastUpdatedAt = now.minus(1, MINUTES),
-          ),
-          aValidReviewScheduleHistory(
-            reference = reviewSchedule3Reference,
-            version = 1,
-            lastUpdatedAt = now.minus(10, MINUTES),
-          ),
-          aValidReviewScheduleHistory(
-            reference = reviewSchedule1Reference,
-            version = 2,
-            lastUpdatedAt = now.minus(365, DAYS),
-          ),
-          aValidReviewScheduleHistory(
-            reference = reviewSchedule2Reference,
-            version = 1,
-            lastUpdatedAt = now.minus(5, DAYS),
-          ),
-          aValidReviewScheduleHistory(
-            reference = reviewSchedule1Reference,
-            version = 1,
-            lastUpdatedAt = now.minus(400, DAYS),
-          ),
-          aValidReviewScheduleHistory(
-            reference = reviewSchedule2Reference,
-            version = 2,
-            lastUpdatedAt = now.minus(4, DAYS),
-          ),
-        ),
-      )
-
-      val expected = listOf(
-        // Review schedule 3 first as it's updated dates are the most recent
-        "Reference: $reviewSchedule3Reference; Version: 2",
-        "Reference: $reviewSchedule3Reference; Version: 1",
-        // Review schedule 2 next
-        "Reference: $reviewSchedule2Reference; Version: 2",
-        "Reference: $reviewSchedule2Reference; Version: 1",
-        // Review schedule 1 last as it's updated dates are the earliest
-        "Reference: $reviewSchedule1Reference; Version: 2",
-        "Reference: $reviewSchedule1Reference; Version: 1",
-      )
-
-      // When
-      val actual = service.getReviewSchedulesForPrisoner(PRISON_NUMBER)
-
-      // Then
-      assertThat(actual.map { "Reference: ${it.reference}; Version: ${it.version}" }).isEqualTo(expected)
-      verify(reviewSchedulePersistenceAdapter).getReviewScheduleHistory(PRISON_NUMBER)
-    }
-
-    @Test
-    fun `should get review schedules given prisoner has no previous review schedules`() {
-      // Given
-      given(reviewSchedulePersistenceAdapter.getReviewScheduleHistory(any())).willReturn(emptyList())
-
-      // When
-      val actual = service.getReviewSchedulesForPrisoner(PRISON_NUMBER)
-
-      // Then
-      assertThat(actual).isEmpty()
-      verify(reviewSchedulePersistenceAdapter).getReviewScheduleHistory(PRISON_NUMBER)
     }
   }
 }
